Consolidate models to basic external router module
For this "external" router, we'll use it to demonstrate basic path
operations, and keeping the models co-located with the path operation
functions makes it a little easier to demonstrate. By getting rid of the
models module, we also make space for a better architecture for
separating internal and API models.