Initial demo app around Astroplan
This sets up a more realistic application within the bootcamp app that's
built around Astroplan and its Observer as the domain model. This
implements the full set of architectural layers we use in our real API
apps:

- Storage layer
- Service layer
- HTTP path operations and API models
- A Factory pattern for creating services and storage clients
- The RequestContext pattern for creating a factory and providing other
  dependencies to endpoints.
- Error handling for user request arguments

Note on vendoring the data for the sites store:

Because the Astropy astropy.coordinates.EarthLocation includes
duplicates of sites under aliases, we're including the original
site.json file in the package itself so we can use our own constructor
for the observer object. Potentially we could to an httpx request to the
Astropy data server at
http://www.astropy.org/astropy-data/coordinates/sites.json but vendoring
makes things easier for this demo.