Add a lesson on dependencies
This lesson explores custom FastAPI dependencies. We cover the two
styles.

For a functional dependency, we're adding a pagination dependency that
adds query strings parameters for a user to control pagination.

For a class-based singleton dependency, we have a simple example that
initializes a semi-random string to show how that value is persistent
across requests.